requirements.txt
python-dotenv
google.generativeai
streamlit
sentence-transformers==2.2.2
pinecone-client
pytesseract
langchain
groq
gtts
lightrag
langchain-community
google-ai-generativelanguage==0.6.5
google-api-core==2.19.1
google-api-python-client==2.134.0
google-auth==2.30.0
huggingface-hub==0.23.4
keras==3.4.0
langchain-text-splitters==0.2.1
langchainplus-sdk==0.0.20
langsmith==0.1.81
pandas==2.2.2
pillow==10.3.0
pinecone-client==4.1.2
pytesseract==0.3.10
scikit-learn==1.5.1
scipy==1.14.0
sentence-transformers==2.2.2
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
torch==2.3.1
torchvision==0.18.1
easyocr
pypdf